Port to 26.2-snapshot-4
