+++ fabric-renderer-api-v1/src/client/resources/fabric-renderer-api-v1.mixins.json
@@ -14,10 +14,11 @@
     "block.particle.BlockStateModelSetMixin",
     "block.particle.ScreenEffectRendererMixin",
     "block.particle.TerrainParticleMixin",
-    "block.render.BlockFeatureRendererMixin",
+    "block.render.BlockModelFeatureRendererMixin",
     "block.render.BlockModelRenderStateMixin",
     "block.render.BlockStateModelWrapperMixin",
     "block.render.LevelExtractorMixin",
+    "block.render.MovingBlockFeatureRendererMixin",
     "block.render.SectionCompilerMixin",
     "item.CuboidItemModelWrapperMixin",
     "item.ItemStackRenderStateLayerRenderStateMixin",
