+++ gradle.properties
@@ -3,7 +3,7 @@ org.gradle.parallel=true
 org.gradle.configuration-cache=false
 
 version=0.146.1
-minecraft_version=26.2-snapshot-3
+minecraft_version=26.2-snapshot-4
 loader_version=0.18.4
 installer_version=1.0.1
